Fix image upload failing on the release page for large photos
The release page's "Replace image" upload sent the raw, full-resolution
file to /api/release/image. The server caps uploads at ~1.5 MB
(MAX_IMAGE_BASE64_LENGTH), so any normal phone photo or hi-res cover was
rejected with 400 "imageBase64 is too large" and surfaced as "Failed to
upload image."

The add-form scan flow already downscaled images to 1024px JPEG before
uploading, which is why only the release page was affected. Extract that
encoding into a shared encodeImageFile() helper and use it in both places
so uploads stay under the server limit consistently.
